Added function to convert netcdf to CLASS .MET file
function can either take a path to a netcdf file (output, scaled file)
for standalone use, or an open python netcdf4 dataset object (for use
inside the scaling functions to optionally output .MET files).  In future,
it would be nice to keep track of station names in the netcdfs so that the
saved met files have better identification than just "site_1", etc.

Also, the date attribute for MERRA and JRA seems to be screwed up for the final
(scaled) files.