fixed bug where time units would be wonky (Issue #14)
The time units were assigned based on the name of the output file. If
the output file name didn't contain 'jra' in the right position, the
units would not be assigned proprly.

Fixed this by changing the function ScaledFileOpen in generic.py to require
a keyword argument specifying the time units in the output file.
This is now explicitly declared in each of the scaling files.