moved topography download (era_to.nc) block to before other data download
Previously, if the download failed halfway through, the directory would
be missing the era_to.nc file, which is required for interpolation.
Now the era_to.nc is downloaded before any other files, so even if the download
fails later, the data can be easily used.

This has the consequence of adding time at the beginning of each download ,
but we can add an argument that allows you to skip the era_ton.nc file if it already
exists (and the areal extent hasn't changed).