moved time-invariant download block to before time-dependent variable download
Previously, if a download crashed partway through, the time-invariant merra_sc file
would be unavailable and therefore the data could not be scaled any further.
By moving this to the front, it downloads first, and therefore partly-completed downloads
can be used while the remainder downlads (if you change the parameter file to pick up
where the last download finished off)

Also did some code cleanup in the MERRAdownlad.retrieve() method. There were a few
redundancies in the definition of the get_variables variable.