rename globsim.py to globsim_main.py
The globsim.py file has the same name as the globsim module (globsim). Therefore, when you want to import a specific function from a file in the directory (say, from globsim.generic import ParameterIO), you get an error because it is trying to import from globsim.py and not the main module.