Changed import statements to look in local module directory
Changed all imports that import from globsim files to ensure that they
look in the right place.  Previously, the scripts used:
"from generic import ParameteRIO", but if the user wasn't
in the globsim directory, you would get an error because python
wouldn't know where to look for the files.  .  Import statements
for files wtihin the globsim module now are preceded by a periodso that they are taken from
inside the module.