Can now download ERA5 using CDS API as well as ECMWF API
 Access through ECMWF Web-API to dataset ERA5 and C3S Seasonal will be phased out.
The CDS API is the new preferred method to access the ERA5 data.  It can access
the data on the MARS server (like ecmwfapi) as well as on the CDS server.

The MARS server requests through CDSAPI can use the same dictionary queries
as those used by ECMWFAPI.  There are a few changes to the queries to CDS,
and these have been implemented as a 'patch' to the final query in
ERA5generic.download().

The default api / storage (ECMWFAPI / ECMWF-MARS) is currently still hard-coded
into the class.  TO change it, just change the class attributes server and storage.

Future updates should include the ability to specify the desired api/server more easily
and eventually transition to supporting only CDS