Added easier re-start function to merra
Before downloading, helper function get_begin_date() looks
in the directory for existing files and makes an inventory using
MFDataset().  Downloading begins on the day after the latest 'complete'
day.  In this case, 'complete' means all 3 downloads (sf, sa, pl) are done.

Note that it does not check the size or completeness of the download. Hopefully
this does not cause problems (i.e. if the download crashes halfway through).
Fortunately this hasn't been a big deal with MERRRA but should be re-evaluated if
applied to ERA5 (unless the download problem gets sorted out).

This helper function get_begin_date() can be used to improve re-start functions
for other datasets as well