add timeout and start to remove ECMWF server
added timeout to server access and download to combat the issue
where downloads hang.  Currently the timeout applies to the
entire request/process/download chain, because to target
only the download part would require modification of the CDS API
which is impractical for redistribution

Started to remove flow control that deals with the server and storage selection
since ECMWF access has been totally phased out.