transpose interpolated field before filling empty netcdf
Merra 

When running globsim with the newer version of `netCDF4==1.5.3` (might actually be caused by a newer version of numpy... I'm not sure), the following error appears:

```
AttributeError: Incompatible shape for in-place modification. Use `.reshape()` to make a copy with the desired shape.
```
My suspicion is that the old version allowed you to be lazy, and if your array had the right number of elements in it, then it would just fill up the netcdf array, even if they weren't the same shape. This means that the interpolated arrays aren't filled in properly, and the station data for MERRA is actually a mix of data from each spot. 

 I tested this, and found that if you use duplicate coordinates in siteslist, the data series aren't the same.

Other interpolaters (JRA, ERA5) have this transpose already.  somehow MERRA was missed.

A good test for data validity would be to run the interpolate with identical siteslist.csv coordinates and ensure the data series are identical.