fix bug with precip scaling in JRA55
previously, code was modified to harmonize JRA55 and JRA3Q values
before running them through the scaling kernel.  But the JRA3Q adjustedment
was being applied to jRA55 so the precip values were too big.

Changed it so the inverse scale (1/(24*3600) is applied to jRA55 and the
scalign kernel expects mm/s